Move logging to parent class
